requirements.txt
alabaster==0.7.16
amqp==5.2.0
anyio==4.4.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
billiard==4.2.0
black==24.4.2
boto3==1.35.17
botocore==1.35.17
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.1
cssbeautifier==1.15.1
decorator==5.1.1
distlib==0.3.8
django-stubs-ext==5.0.4
django-timezone-field==7.0
docutils==0.21.2
EditorConfig==0.12.4
exceptiongroup==1.2.2
executing==2.1.0
Faker==28.4.1
fido2==1.1.3
filelock==3.16.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
humanize==4.10.0
identify==2.6.0
idna==3.8
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg-c==3.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.38.5
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240907
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.26.4
wcwidth==0.2.13
websockets==13.0.1
